refact: use native Promises instead of Bluebird
Bluebird is definitely still faster, but native Promises have made great strides in closing the gap. As well, the integration of native promises for tracing and debuggability are quite desirable, so continuing to transition libraries over (especially as native Promise is now widely available on the node.js versions we expect people to be running.